modified profile resource

+++ src/main/java/org/mk/learning/messenger/model/Profile.java
@@ -17,7 +17,7 @@ public class Profile {
 		
 	}
 	
-	public Profile(long id, String profileName, String firstName, String lastName) {
+	public Profile(int id, String profileName, String firstName, String lastName) {
 		this.id=id;
 		this.profileName=profileName;
 		this.firstName=firstName;
